requirements.txt
absl-py==0.11.0
alibi==0.5.5
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
BHClustering==0.0.3
bleach==3.1.0
blis==0.4.1
bokeh==2.2.3
boto3==1.15.4
botocore==1.18.4
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
cmake==3.18.2.post1
colorama==0.4.1
colorcet==2.0.6
conda==4.7.12
conda-package-handling==1.3.11
cryptography==2.7
cycler==0.10.0
cymem==2.0.5
Cython==0.29.16
dask==2021.2.0
dataclasses==0.6
datashader==0.12.0
datashape==0.5.2
decorator==4.4.1
defusedxml==0.6.0
dice-ml==0.4
distlib==0.3.0
distributed==2021.2.0
drive-cli==2.1.0
entrypoints==0.3
fasttext==0.9.2
ffmpeg==1.4
filelock==3.0.12
fire==0.3.1
flatbuffers==1.12
flatten-json==0.1.13
fsspec==0.8.3
future==0.18.2
gast==0.3.3
glfw==1.11.0
google-api-python-client==1.7.8
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
holoviews==1.14.1
httplib2==0.12.1
idna==2.8
imageio==2.6.1
importlib-metadata==1.4.0
ipykernel==5.3.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jax==0.2.9
jaxlib==0.1.61+cuda101
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.17.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
libarchive-c==2.8
lime==0.2.0.1
llvmlite==0.35.0
lmdb==1.0.0
locket==0.2.1
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.1.0
msgpack==1.0.2
MulticoreTSNE==0.1
multipledispatch==0.6.0
murmurhash==1.0.5
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.4
nltk==3.5
notebook==6.0.2
numba==0.52.0
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
openai==0.3.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.5
pandas-path==0.1.2
pandocfilters==1.4.2
panel==0.10.3
param==1.10.1
parso==0.5.2
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pick==0.6.4
pickleshare==0.7.5
Pillow==7.2.0
plac==0.9.6
preshed==3.0.5
prettytable==0.7.2
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.13.0
protobuf3-to-dict==0.1.5
psutil==5.8.0
ptpython==3.0.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycosat==0.6.3
pycparser==2.19
pyct==0.4.8
pyDOE2==1.3.0
pyfiglet==0.8.post1
pyglet==1.5.0
Pygments==2.5.2
pynndescent==0.5.2
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.0
python-dateutil==2.8.1
pytorch-lightning==0.10.0
pytz==2021.3
pyviz-comms==2.0.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qbstyles==0.1.3
qtconsole==4.6.0
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.3.3
sacremoses==0.0.43
sagemaker==2.10.0
scikit-image==0.16.2
scikit-learn==0.22.1
scikit-video==1.1.11
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.91
shap==0.39.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smdebug==0.9.3
smdebug-rulesconfig==0.1.5
sortedcontainers==2.3.0
soupsieve==2.1
spacy==2.2.0
spacy-lookups-data==0.3.2
srsly==1.0.5
statsmodels==0.11.1
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.0.0.dev2019092611
tfrecord==1.11
thinc==7.1.1
threadpoolctl==3.0.0
tokenizers==0.10.1
toolz==0.11.1
torch==1.8.1
torchvision==0.9.1
tornado==6.0.3
tqdm==4.62.3
traitlets==4.3.3
transformers==4.3.2
typing-extensions==3.7.4.3
umap==0.1.1
umap-learn==0.5.1
uritemplate==3.0.1
urllib3==1.24.2
Utils-Config==1.0.0
virtualenv==20.0.16
wasabi==0.8.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.16.2
zict==2.0.0
zipp==2.0.0
